backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pinecone-client==3.0.0
openai==1.3.0
langchain==0.1.0
langchain-community==0.0.8
llama-cpp-python==0.2.26
requests==2.31.0
numpy==1.26.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
pydantic-settings==2.1.0
cors==1.0.1
